AutogenBenchmarking/pyproject.toml
[tool.poetry]
name = "ag-benchmarking"
version = "0.1.0"
description = "A comprehensive benchmarking system for AutoGen agents"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ag_benchmarking"}]

[tool.poetry.dependencies]
python = "^3.8"
autogen-agentchat = ">=0.2.0"
fastapi = ">=0.68.0"
uvicorn = ">=0.15.0"
pydantic = ">=1.8.0"
psutil = ">=5.8.0"
click = ">=8.0.0"
pyyaml = ">=5.4.0"
websockets = ">=10.0"
python-dotenv = ">=0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-asyncio = ">=0.16.0"
pytest-cov = ">=2.12.0"
flake8 = ">=3.9.0"
black = ">=21.5b2"
isort = ">=5.9.0"
mypy = ">=0.910"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true